pyproject.toml
[project]
name = "evade84-node"
version = "0.1.0"
description = "evade84 node"
authors = [
    {name = "jieggii", email = "jieggii@pm.me"},
]
dependencies = [
    "fastapi>=0.78.0",
    "uvicorn[standard]>=0.17.6",
    "beanie>=1.11.1",
    "motor>=2.5.1",
    "betterconf>=2.6.1",
    "argon2-cffi>=21.3.0",
    "loguru>=0.6.0",
    "pycryptodome>=3.14.1",
    "shortuuid>=1.0.9",
]
requires-python = ">=3.10"
license = {text = "MIT"}
[project.optional-dependencies]

[tool]
[tool.pdm]

[tool.pdm.scripts]
start.cmd = "uvicorn node.main:app --reload --reload-dir ./node/"
start.env_file = "debug.env"

fmt.shell = "isort ./node/ && black ./node/"
lint = "flake8 ./node/"

[tool.pdm.dev-dependencies]
dev = [
    "isort>=5.10.1",
    "black>=22.3.0",
    "flake8>=4.0.1",
]

[tool.black]
line-length = 105
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 105
src_paths = ["isort", "test"]

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"
